+++ bin/test-workflow/5_app_namespace_prep.sh
@@ -7,11 +7,14 @@ set_namespace default
 
 # Prepare a given namespace with a subset of credentials from the golden configmap
 announce "Installing application namespace prep chart"
-pushd helm/application-namespace-prep
-    helm uninstall namespace-prep -n "$TEST_APP_NAMESPACE_NAME"
+pushd $(dirname "$0")/../../helm/application-namespace-prep > /dev/null
+    if [ "$(helm list -q -n $TEST_APP_NAMESPACE_NAME | grep "^namespace-prep$")" = "namespace-prep" ]; then
+        helm uninstall namespace-prep -n "$TEST_APP_NAMESPACE_NAME"
+    fi
 
     # Namespace $TEST_APP_NAMESPACE_NAME will be created if it does not exist
-    helm install namespace-prep . -n "$TEST_APP_NAMESPACE_NAME"  --wait \
-        --set authnK8s.goldenConfigMap="$TEST_APP_NAMESPACE_NAME" \
+    helm install namespace-prep . -n "$TEST_APP_NAMESPACE_NAME"  --debug --wait \
+        --create-namespace \
+        --set authnK8s.goldenConfigMap="authn-k8s-configmap" \
         --set authnK8s.namespace="$CONJUR_NAMESPACE"
-popd
+popd > /dev/null
